Reduced memory footprint of ModuleExtClock and ModuleClock

+++ ModuleClock.h
@@ -50,8 +50,8 @@ class ModuleClock : public Module
     uint32_t rate;
     uint32_t counter;
     uint32_t bpm;
-    uint32_t bpm_ppqn[255];
-    uint32_t bpm_half_ppqn[255];
+    uint16_t bpm_ppqn[255];
+    uint16_t bpm_half_ppqn[255];
 };
 
 #endif
