+++ ModuleExtClock.cpp
@@ -16,7 +16,7 @@ ModuleExtClock::ModuleExtClock(uint32_t bpm, int clock_division)
   // 'bpm' the beats per minute of the internal clock
   this->bpm = constrain(bpm, 0, 254);
   
-  for(uint32_t bpm_i=0; bpm_i < 255; bpm_i++)
+  for(uint16_t bpm_i=0; bpm_i < 255; bpm_i++)
   {
     bpm_ppqn[bpm_i] = ((float)(60.0 * SAMPLE_RATE_FLOAT * clock_division)/((float)bpm_i * 96.0));
     bpm_half_ppqn[bpm_i] = bpm_ppqn[bpm_i] / 2;
@@ -30,7 +30,7 @@ ModuleExtClock::ModuleExtClock(uint32_t bpm)
   this->ext_clock_counter = 176400;
   this->bpm = constrain(bpm, 0, 254);
 
-  for(uint32_t bpm_i=0; bpm_i < 255; bpm_i++)
+  for(uint16_t bpm_i=0; bpm_i < 255; bpm_i++)
   {
     bpm_ppqn[bpm_i] = ((float)(60.0 * SAMPLE_RATE_FLOAT)/((float)bpm_i * 96.0));
     bpm_half_ppqn[bpm_i] = bpm_ppqn[bpm_i] / 2;
